# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: CAA Compliance Engine — Internal API
  version: 0.1.0
  description: >
    Internal contract between the PHP/Laravel public site & Client Portal
    and the Java/Spring Boot Compliance Engine (Section 2 of
    CAA-Website-Rebuild-Architecture.md). PHP calls this API; it never
    talks to the database directly. Not public-facing — no external auth
    scheme is specified here, only the shape of the contract.
servers:
  - url: http://compliance-engine.internal:8080/api/v1
security:
  - basicAuth: []
...
paths:
...
  /certification-runs:
    post:
      tags: [certifications]
      summary: >
        Record a completed CertificationRun and its Observations. Scores
        per Section 3b's two-condition rule (37 cumulative cap per color,
        independently; 20% failed-reading threshold). point_count must be
        50 (initial) or 25 (White-only split retake) — Black can never be
        split. After scoring, determines whether this run resulted in an
        overall pass (standard, or White-split-retake combined with the
        prior run's already-passing Black) — if so, creates a
        Certification and flips the Enrollment to CERTIFIED (Section 4g).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificationRunCreateRequest'
      responses:
        '201':
          description: Created and scored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificationRunCreateResponse'
        '422':
          description: >
            Invalid — e.g. a 25-point run with Black observations attached,
            or a split retake that doesn't chain to a valid prior run.
  /certification-runs/{runId}/split-run-eligibility:
    get:
      tags: [certifications]
      summary: >
        Evaluate split-run eligibility for a completed run (Section 3b).
        White-only — Black failing always returns NOT_ELIGIBLE_BLACK_FAILED
        regardless of White's result.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    enum:
                      - ELIGIBLE_WHITE_RETAKE
                      - NOT_ELIGIBLE_BLACK_FAILED
                      - NOT_ELIGIBLE_BOTH_PASSED
                      - BLOCKED_REGION
                      - BLOCKED_FORMAT

  # ---------------------------------------------------------------------
  # Clients (Section 3, 3c)
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        Section 3: the entire API requires HTTP Basic Auth as of the
        security addition -- previously wide open. Applies globally
        (SecurityConfig's anyRequest().authenticated()), not just to the
        one path that explicitly lists it below; that explicit listing on
        /auth/me is just to call out its special role in the login flow.
...
  schemas:
...
    ObservationInput:
      type: object
      required: [color, pointNumber, trueOpacityValue, studentEstimatedOpacity]
      properties:
        color:
          type: string
          enum: [WHITE, BLACK]
        pointNumber: { type: integer }
        trueOpacityValue:
          type: integer
          description: Always a multiple of 5 (Section 4g).
        studentEstimatedOpacity:
          type: integer
          description: Always a multiple of 5 (Section 4g).
    CertificationRunCreateRequest:
      type: object
      required: [sessionId, enrollmentId, runNumber, pointCount, trailerId, testingSystemId, administeredByStaffId, observations]
      properties:
        sessionId: { type: integer, format: int64 }
        enrollmentId:
          type: integer
          format: int64
          description: >
            Which Enrollment this attempt belongs to. Not stored on
            CertificationRun itself (that only references Session) --
            required here so CertificationDeterminationService knows
            which Enrollment to flip to CERTIFIED on an overall pass.
        runNumber: { type: integer }
        pointCount:
          type: integer
          enum: [25, 50]
        priorRunId:
          type: integer
          format: int64
          nullable: true
          description: Required for a 25-point split retake — must chain to the run it succeeds.
        trailerId: { type: integer, format: int64 }
        testingSystemId: { type: integer, format: int64 }
        administeredByStaffId: { type: integer, format: int64 }
        observations:
          type: array
          items:
            $ref: '#/components/schemas/ObservationInput'
    CertificationRun:
      type: object
      properties:
        id: { type: integer, format: int64 }
        runNumber: { type: integer }
        pointCount: { type: integer }
        whiteCumulativeDeviation: { type: number, format: double, nullable: true }
        whitePass: { type: boolean, nullable: true }
        blackCumulativeDeviation: { type: number, format: double, nullable: true }
        blackPass: { type: boolean, nullable: true }
        inProgress:
          type: boolean
          description: >
            True only during a live test (LiveTestingService), between
            creation and final scoring -- false for every run created
            via the original batch-creation endpoint, which is always
            already-scored at insert time.
    Certification:
      type: object
      properties:
        id: { type: integer, format: int64 }
        issueDate: { type: string, format: date, nullable: true }
        expirationDate: { type: string, format: date, nullable: true }
        pdfCertificateLink: { type: string, nullable: true }
        whiteCumulativeDeviation: { type: number, format: double, nullable: true }
        blackCumulativeDeviation: { type: number, format: double, nullable: true }
        anyFailedReading: { type: boolean }
        passFail: { type: boolean, nullable: true }
    CertificationRunCreateResponse:
      type: object
      properties:
        run: { $ref: '#/components/schemas/CertificationRun' }
        certified:
          type: boolean
          description: True if this run resulted in an overall pass (standard or White-split-retake combination).
        certification:
          $ref: '#/components/schemas/Certification'
...